fix(chip8): cast PC arithmetic to silence -Warith-conversion
GCC/Clang perform integer promotions on uint16_t expressions, so the
address computation `c8->pc + 1` is evaluated as `int` and then implicitly
converted back to `uint16_t` when passed to `memory_read()`. With
`-Warith-conversion` enabled (and promoted to an error via `-Werror`),
this triggers a hard failure:

  conversion from `int` to `uint16_t` may change value [-Werror=arith-conversion]

Cast the final expression to `uint16_t` to make the narrowing explicit and
keep builds clean with the stricter warning set.

+++ src/chip8.c
@@ -9,7 +9,7 @@ static inline uint16_t
 chip8_fetch(const Chip8 *c8)
 {
 	const uint8_t hi = memory_read(&c8->mem, c8->pc);
-	const uint8_t lo = memory_read(&c8->mem, c8->pc + 1);
+	const uint8_t lo = memory_read(&c8->mem, (uint16_t)(c8->pc + 1u));
 	return (uint16_t)((hi << 8) | lo);
 }
